+++ R/distantia_time_delay.R
@@ -1,17 +1,23 @@
-#' Time Shift Between Time Series
+#' Time Delay Between Time Series
 #'
 #' @description
-#' This function computes an approximation to the time-shift between pairs of time series as the absolute time difference between pairs of observations in the time series *x* and *y* connected by the dynamic time warping path.
+#' This function computes an approximation to the time-delay between pairs of time series as the difference between observations connected by the dynamic time warping path.
 #'
-#' If the time series are long enough, the extremes of the warping path are trimmed (5% of the total path length each) to avoid artifacts due to early misalignments.
+#' Given a pair of time series `x` and `y`, and the time of their samples in the dynamic time warping `time(x)` and `time(y)`, when the argument `directional` is `TRUE`, the time delay is computed as follows:
+#'\itemize{
+#' \item Time delay from `x` to `y`: `time(y) - time(x)`.
+#' \item Time delay from `y` to `x`: `time(x) - time(y)`
+#'}
+#' In such case, two rows per pair of time series are returend. Otherwise, the time delay is computed as `abs(time(y) - time(x))`, and only one row per pair of time series is returned.
 #'
-#' It returns a data frame with the modal, mean, median, minimum, maximum, quantiles 0.25 and 0.75, and standard deviation. The modal and the median are the most generally accurate time-shift descriptors.
+#' If the time series have more than 30 observations, a 5% of the cases are omitted at each extreme of the warping path to avoid overestimating time delays due to early misalignments.
+#'
+#' The function returns a data frame with the names of the time series in the columns *x* and *y*, and the stats of the time-delay. The modal and the median are the most generally accurate time-delay metrics
 #'
 #' This function requires scaled and detrended time series. Still, it might yield non-sensical results in case of degenerate warping paths. Plotting dubious results with [distantia_dtw_plot())] is always a good approach to identify these cases.
 #'
-#'
 #' @inheritParams distantia
-#' @param two_way (optional, logical) If TRUE, the time shift between the time series pairs *y* and *x* is added to the results
+#' @param directional (optional, logical) If TRUE, a directional time-delay is computed as `x to y` and `y to x`, resulting in two rows per pair of time series. Otherwise, the absolute magnitude of de delay between `x` and `y` is returned a a single row per pair. Default: TRUE
 #'
 #' @return data frame
 #' @export
@@ -46,19 +52,16 @@
 #' #compute shifts
 #' df_shift <- distantia_time_delay(
 #'   tsl = tsl,
-#'   two_way = TRUE
+#'   directional = TRUE
 #' )
 #'
 #' df_shift
-#' #positive shift values indicate
-#' #that the samples in Kinshasa
-#' #are aligned with older samples in London.
 #' @family distantia_support
 distantia_time_delay <- function(
     tsl = NULL,
     distance = "euclidean",
     bandwidth = 1,
-    two_way = FALSE
+    directional = FALSE
 ){
 
   #check input arguments
@@ -79,7 +82,7 @@ distantia_time_delay <- function(
   )
 
   #function to compute modal of differences
-  shift_modal <- function(x) {
+  delay_modal <- function(x) {
     ux <- unique(na.omit(x))
     y <- ux[which.max(tabulate(match(x, ux)))][1]
     class(y) <- class(x)
@@ -92,7 +95,7 @@ distantia_time_delay <- function(
   p <- progressr::progressor(along = iterations)
 
   #iterate over pairs of time series
-  df_shift <- foreach::foreach(
+  df_delay <- foreach::foreach(
     i = iterations,
     .combine = "rbind",
     .errorhandling = "pass",
@@ -127,38 +130,39 @@ distantia_time_delay <- function(
     #compute n to assess if we apply padding
     n <- nrow(cost_path.i) - (padding * 2)
 
+    #apply padding when 30 or more cases
     if(n >= 30){
 
-      cost_path.i <- cost_path.i[padding:(nrow(cost_path.i)-padding),]
+      cost_path.i <- cost_path.i[padding:(nrow(cost_path.i) - padding),]
 
     }
 
     #check time series units
-    shift.time.i.units.x <- zoo_time(x = tsl[[df.i[["x"]]]])$units
+    delay.units.x <- zoo_time(x = tsl[[df.i[["x"]]]])$units
+    delay.units.y <- zoo_time(x = tsl[[df.i[["y"]]]])$units
 
-    shift.time.i.units.y <- zoo_time(x = tsl[[df.i[["y"]]]])$units
-
-    if(shift.time.i.units.x != shift.time.i.units.y){
+    #compare by sample if units do not match
+    if(delay.units.x != delay.units.y){
 
       warning(
         "distantia::distantia_time_delay(): time series '",
         df[["x"]], "' and '", df[["y"]], "' have different time units (",
-        shift.time.i.units.x,
+        delay.units.x,
         " vs ",
-        shift.time.i.units.y,
-        ". Computing time shift with units 'samples'."
+        delay.units.y,
+        ". Computing time delay with units 'samples'."
       )
 
-      shift.time.i.units <- "samples"
+      #change default units
+      delay.units <- "samples"
 
-      #time difference
-      shift.time.i <- cost_path.i[["x"]] - cost_path.i[["y"]]
+      #sample columns
+      colname_x <- "x"
+      colname_y <- "y"
 
     } else {
 
-      shift.time.i.units <- shift.time.i.units.x
-
-      #add time
+      #create time columns
       cost_path.i[["x_time"]] <- zoo::index(
         x = tsl[[df.i[["x"]]]]
       )[cost_path.i[["x"]]]
@@ -167,70 +171,108 @@ distantia_time_delay <- function(
         x = tsl[[df.i[["y"]]]]
       )[cost_path.i[["y"]]]
 
-      #time difference
-      shift.time.i <- cost_path.i[["x_time"]] - cost_path.i[["y_time"]]
+      #default delay units
+      delay.units <- delay.units.x
+
+      #default time column names
+      colname_x <- "x_time"
+      colname_y <- "y_time"
 
     }
 
-    #convert to numeric to remove units
-    shift.time.i <- as.numeric(shift.time.i)
+    #time difference from x to y
+    if(directional == FALSE){
 
-    df.i[["units"]] <- shift.time.i.units
+      delay.x_to_y <- abs(as.numeric(cost_path.i[[colname_x]] - cost_path.i[[colname_y]]))
+
+    } else {
+
+      delay.x_to_y <- as.numeric(cost_path.i[[colname_y]] - cost_path.i[[colname_x]])
+
+      delay.y_to_x <- as.numeric(cost_path.i[[colname_x]] - cost_path.i[[colname_y]])
+
+    }
+
+    #add units
+    df.i[["units"]] <- delay.units
 
     df.i[["min"]] <- min(
-      x = shift.time.i,
+      x = delay.x_to_y,
       na.rm = TRUE
     )
 
     df.i[["q1"]] <- stats::quantile(
-      x = shift.time.i,
-      probs = 0.25,
+      x = delay.x_to_y,
+      na.rm = TRUE,
+      probs = 0.25
+    )
+
+    #compute stats x to y
+    df.i[["median"]] <- stats::median(
+      x = delay.x_to_y,
       na.rm = TRUE
     )
 
-    df.i[["median"]] <- stats::median(
-      x = shift.time.i,
-      na.rm = TRUE
-    )
+    df.i[["modal"]] <- delay_modal(x = delay.x_to_y)
 
     df.i[["mean"]] <- mean(
-      x = shift.time.i,
+      x = delay.x_to_y,
       na.rm = TRUE
     )
 
-    df.i[["modal"]] <- shift_modal(x = shift.time.i)
-
-    df.i[["q3"]] <- stats::quantile(
-      x = shift.time.i,
-      probs = 0.75,
-      na.rm = TRUE
+    df.i[["q3"]][1] <- stats::quantile(
+      x = delay.x_to_y,
+      na.rm = TRUE,
+      probs = 0.75
     )
 
     df.i[["max"]] <- max(
-      x = shift.time.i,
+      x = delay.x_to_y,
       na.rm = TRUE
     )
 
-    df.i[["sd"]] <- stats::sd(
-      x = shift.time.i,
-      na.rm = TRUE
-    )
-
-    #add tow way df
-    if(two_way == TRUE){
+    #compute stats y to x
+    if(directional == TRUE){
 
       df.j <- df.i
+      df.j$x <- df.i$y
+      df.j$y <- df.i$x
 
-      df.j[["x"]] <- df.i[["y"]]
-      df.j[["y"]] <- df.i[["x"]]
 
-      df.j[["min"]] <- - df.i[["min"]]
-      df.j[["q1"]] <- - df.i[["q1"]]
-      df.j[["median"]] <- - df.i[["median"]]
-      df.j[["mean"]] <- - df.i[["mean"]]
-      df.j[["modal"]] <- - df.i[["modal"]]
-      df.j[["q3"]] <- - df.i[["q3"]]
-      df.j[["max"]] <- - df.i[["max"]]
+      df.j[["min"]] <- min(
+        x = delay.y_to_x,
+        na.rm = TRUE
+      )
+
+      df.j[["q1"]] <- stats::quantile(
+        x = delay.y_to_x,
+        na.rm = TRUE,
+        probs = 0.25
+      )
+
+      df.j[["median"]] <- stats::median(
+        x = delay.y_to_x,
+        na.rm = TRUE
+      )
+
+      df.j[["modal"]] <- delay_modal(x = delay.y_to_x)
+
+
+      df.j[["mean"]] <- mean(
+        x = delay.y_to_x,
+        na.rm = TRUE
+      )
+
+      df.j[["q3"]] <- stats::quantile(
+        x = delay.y_to_x,
+        na.rm = TRUE,
+        probs = 0.75
+      )
+
+      df.j[["max"]] <- max(
+        x = delay.y_to_x,
+        na.rm = TRUE
+      )
 
       df.i <- rbind(
         df.i,
@@ -243,12 +285,14 @@ distantia_time_delay <- function(
 
   }
 
+  rownames(df_delay) <- NULL
+
   #add type
   attr(
-    x = df_shift,
+    x = df_delay,
     which = "type"
   ) <- "time_delay_df"
 
-  df_shift
+  df_delay
 
 }